refactor(frontend): extract SourceRowActions; SourceRow now 149 lines (SPEC-PORTAL-SOURCES-RENAME-001 AC-4)
The earlier 297-line -sources-row.tsx mixed three concerns in one
component: row layout, inline rename overlay, and the 7-button action
cluster. Split the cluster into -sources-row-actions.tsx; SourceRow
now only orchestrates name + status + drill-down + rename overlay,
delegating the buttons.

Line counts after split (all under 200, AC-4 met):
  sources.tsx                152 lines
  -sources-row.tsx           149 lines   (was 297)
  -sources-row-actions.tsx   175 lines   (new)
  -sources-hooks.ts          175 lines
  -sources-content.tsx       131 lines
  -sources-helpers.tsx        86 lines
  -sources-actionbar.tsx      72 lines
  -sources-types.ts           48 lines

The SourceRowActions component owns its own mutations
(useSourceSync / useSourceDelete / useSourceReauth) — passed nothing
but `source` + `kbSlug`. The parent SourceRow keeps
useSourceRename because the rename Save/Cancel overlay is
state-coupled with the InlineEdit primitive (shared draftName).

Verified:
  tsc --noEmit: exit 0
  vitest run src/routes/app/knowledge: 41/41 pass (4 hook tests + 37 KB tests)
